Split check_internal command into smaller pieces, some associated cleanup
