+++ src/commands/check_internal/error.rs
@@ -1,0 +1,22 @@
+use thiserror::Error;
+
+use crate::exclusion;
+use crate::filesystem as fs;
+use crate::interfaces::error::InterfaceError;
+use crate::modules;
+
+#[derive(Error, Debug)]
+pub enum CheckError {
+    #[error("The path {0} is not a valid directory.")]
+    InvalidDirectory(String),
+    #[error("Filesystem error: {0}")]
+    Filesystem(#[from] fs::FileSystemError),
+    #[error("Module tree error: {0}")]
+    ModuleTree(#[from] modules::error::ModuleTreeError),
+    #[error("Exclusion error: {0}")]
+    Exclusion(#[from] exclusion::PathExclusionError),
+    #[error("Interface error: {0}")]
+    Interface(#[from] InterfaceError),
+    #[error("Operation cancelled by user")]
+    Interrupt,
+}
