--- src/tests/lsp_server.rs
@@ -1,7 +1,0 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn test_server_initialization() {
-        // TODO: Difficult to test due to thread spawning, need for stdio connection
-    }
-}
