+++ src/tests/mod.rs
@@ -1,5 +1,4 @@
 pub mod check_internal;
-pub mod lsp_server;
 pub mod module;
 pub mod test;
 #[cfg(test)]
